refactoring mumu class, rewriting calculateHTLT func, moving histogramClass struct to header
